+++ src/pages/transferWritingPricePage/agreementSection/AgreementSection.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from "react";
-import CheckBoxSection from "../checkBox/CheckBoxSection";
 
 import * as S from "./AgreementSection.style";
+import CheckBoxSection from "../checkBox/CheckBoxSection";
 
 interface AgreementProps {
   opt1: boolean;
