+++ src/pages/transferWritingPricePage/enterAccountInfo/EnterAccountInfo.tsx
@@ -1,5 +1,3 @@
-import { BANK_LIST } from "@/constants/bank";
-import { PATH } from "@/constants/path";
 import usePreventLeave from "@hooks/common/usePreventLeave";
 import useToastConfig from "@hooks/common/useToastConfig";
 import { Nullable } from "@type/nullable";
@@ -8,6 +6,9 @@ import { useEffect, useRef, useState } from "react";
 import { useLocation } from "react-router-dom";
 
 import * as S from "./EnterAccountInfo.style";
+
+import { BANK_LIST } from "@/constants/bank";
+import { PATH } from "@/constants/path";
 
 const EnterAccountInfo = ({
   bank,
